tests/performance/wifi/mqtt/normal/receiver/requirements.txt
numpy==1.26.3
paho-mqtt==1.6.1